feat: KeyboardStyle 함수 컴퍼넌트화 #63
